Move CLI nonce args to clap-utils
