+++ cli/src/offline/blockhash_query.rs
@@ -1,4 +1,5 @@
 use super::*;
+use solana_clap_utils::nonce::*;
 use solana_client::nonce_utils;
 use solana_sdk::commitment_config::CommitmentConfig;
 
@@ -76,7 +77,7 @@ impl BlockhashQuery {
     pub fn new_from_matches(matches: &ArgMatches<'_>) -> Self {
         let blockhash = value_of(matches, BLOCKHASH_ARG.name);
         let sign_only = matches.is_present(SIGN_ONLY_ARG.name);
-        let nonce_account = pubkey_of(matches, nonce::NONCE_ARG.name);
+        let nonce_account = pubkey_of(matches, NONCE_ARG.name);
         BlockhashQuery::new(blockhash, sign_only, nonce_account)
     }
 
@@ -109,7 +110,7 @@ impl Default for BlockhashQuery {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::{nonce::nonce_arg, offline::blockhash_query::BlockhashQuery};
+    use crate::offline::blockhash_query::BlockhashQuery;
     use clap::App;
     use serde_json::{self, json, Value};
     use solana_account_decoder::{UiAccount, UiAccountEncoding};
@@ -172,9 +173,7 @@ mod tests {
 
     #[test]
     fn test_blockhash_query_new_from_matches_ok() {
-        let test_commands = App::new("blockhash_query_test")
-            .arg(nonce_arg())
-            .offline_args();
+        let test_commands = App::new("blockhash_query_test").nonce_args().offline_args();
         let blockhash = hash(&[1u8]);
         let blockhash_string = blockhash.to_string();
